Reorganized and optimized actors in the load testing tool #1427
