+++ load-testing/src/main/resources/application.conf
@@ -1,0 +1,8 @@
+akka {
+    http {
+        host-connection-pool {
+            max-connections = 2048
+            max-open-requests = 2048
+        }
+    }
+}
