--- load-testing/src/main/scala/co/topl/loadtesting/ToStatisticsCsvLog.scala
@@ -1,8 +1,0 @@
-package co.topl.loadtesting
-
-import simulacrum.typeclass
-
-@typeclass
-trait ToStatisticsCsvLog[T] {
-  def toLog(t: T): String
-}
